turn resolveName into separate module
